# Conflicts:
#	backend/src/modules/teamMemberships/teamMembershipService.js
#	frontend/src/pages/administration/AdministrationMenu.js

+++ database/sql/ddl.sql
@@ -102,7 +102,7 @@ CREATE TABLE `events` (
     `id_assistance1` int,
     `id_assistance2` int,
     `minute` int,
-    `value` tinyint,
+    `value` int,
     `host` boolean NOT NULL
 );
